requirements.txt
bcrypt==3.2.2
cffi==1.15.1
cryptography==37.0.4
flake8==5.0.4
future==0.18.2
mccabe==0.7.0
netmiko==4.1.2
ntc-templates==3.0.0
paramiko==2.11.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==0.20.0
scp==0.14.4
six==1.16.0
tabulate==0.8.10
tenacity==8.0.1
textfsm==1.1.2
